fir_auth_2fa/requirements.txt
django-otp-yubikey
django-two-factor-auth[phonenumbers]
django-two-factor-auth[webauthn]
